[BE] Job comments https://trello.com/c/vRfRn8EZ/85-be-job-comments
- Moved addComment & editComment to CommentTransformer.
- Moved the comment endpoints to a separate controller CommentController.
- When deleting an user his comments remain actual with user_id equal null
- Removed the bidirectional relation between id - parent
