+++ src/Entity/User.php
@@ -194,9 +194,8 @@ class User implements UserInterface
      * @ORM\OneToOne(targetEntity="Image")
      * @ORM\JoinColumn()
      * @Serializer\Expose()
-     * @Groups({"UserDetail", "UserList", "ActivityList", "ActivityDetails"})
-     * @SWG\Property(ref="#/definitions/Image")
      * @Groups({"UserDetail", "UserList", "ActivityList", "ActivityDetails", "Comment"})
+     * @SWG\Property(ref="#/definitions/Image")
      */
     private $avatar;
